requirements.txt
boto3==1.18.47
celery==5.1.2
fastapi==0.66.0
flower==1.0.0
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2021.1
pyvips==2.1.15
redis==3.2.1
uvicorn==0.14.0
streaming-form-data
hurry.filesize==0.9
pyjwt
orjson
httpx
gunicorn
requests
